Refactor: move getDatabase from app to menu
